requirements.txt

### requirements.txt

Pillow
